rerun of everything after bug fix in "order" in data
There was a bug in preprocessing code that made the "order" column incorrect for some participants. It has been fixed and this is a commit that runs everything with the new data.

One additional change - countries in SI Fig 9 are now sorted alphabetically within the count

All Figs were run in the FMP_github venv